Reorganization for lightcone
