+++ python/opencosmo/collection/lightcone/__init__.py
@@ -1,4 +1,9 @@
+from importlib import import_module
+
 from .healpix_map import HealpixMap
 from .lightcone import Lightcone
+
+# register plugins
+import_module(f"{__name__}.plugins")
 
 __all__ = ["Lightcone", "HealpixMap"]
