+++ python/opencosmo/plugins/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import PluginSpec, PluginType, register_plugin
+
+__all__ = ["register_plugin", "PluginSpec", "PluginType"]
